+++ release_notes.md
@@ -6,6 +6,7 @@ short summary of each release, with version and release date.
 work in progress.
 this work is available on "master" branch, but has not yet been released.
 
+  * `But::Optional` now has relational operators defined.
   * `Log::Field::PreciseDT` for nanosecond-precision timestamps in logs.
   * `Optional::emplace(...)` can now be used to conveniently create objects in place, including those that are non-copyable and non-movable.
   * added support for GCC-8 and Clang-6.